added entry bindings api
